Add first example
A first example for how the new build tool would work has been made.
I still haven't figured out:
  - How to hash the tasks nicely in Go. This would be easy in Java
    (thanks to objects.Hash) but needs to be done manually in Go, and
   ideally we don't push the responsibility to the end-user.
  - The example doesn't terminate, we have bugs.

+++ pkg/nbt/nbt.go
@@ -2,7 +2,7 @@ package nbt
 
 type Task interface {
 	/* Returns a hash for this task that can be used to use it in a map. */
-	Hash() int
+	Hash() uintptr
 	/* Returns true if this task matches the given task, false otherwise. */
 	Matches(Task) bool
 	Perform(h Handler)
